# hand encoding json is fraught with peril, just use stdlib, it knows how.