fixes #241 and I also am assuming any other case of json encoded as string inside json.