@lsegal 2015-05-22
It's worth noting that applying this patch makes the marshalling process roughly 20% slower:

```sh
# before
$ go test -bench . ./internal/protocol/json/jsonutil/
PASS
BenchmarkBuildJSON	   50000	     32727 ns/op
BenchmarkStdlibJSON	  100000	     12064 ns/op
ok  	github.com/awslabs/aws-sdk-go/internal/protocol/json/jsonutil	3.516s

# after
$ go test -bench . ./internal/protocol/json/jsonutil/
PASS
BenchmarkBuildJSON	   50000	     39859 ns/op
BenchmarkStdlibJSON	  100000	     12254 ns/op
ok  	github.com/awslabs/aws-sdk-go/internal/protocol/json/jsonutil	3.930s
```

Given that the specification for string is fairly simple (http://www.json.org/), I'm not sure it warrants the overhead.