@sclasen 2015-05-22
Cool, your last commit actually fixes #241 too, so im fine either way.

Might want to have a spec that does some crazy json inside json inside json inside json and have it actually hit some services. Never know what people will try and shove in dynamodb.