@lsegal 2015-05-22
Yea that is a good idea. I'll take a look tomorrow at beefing up the test suite to make sure we're handling cases like yours.